Update funcionando
Agora falta extrair as strings

+++ app/src/main/java/br/edu/ifsp/dmo/diario/ui/listeners/DiaryEntryItemClickListener.kt
@@ -1,7 +1,7 @@
 package br.edu.ifsp.dmo.diario.ui.listeners
 
 interface DiaryEntryItemClickListener {
-    fun clickOpenDiaryEntry(position: Int)
+    fun clickOpenDiaryEntry(position: Long)
 
     fun clickDeleteDiaryEntry(position: Long)
 }
